+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * Expose the `poll_connection`, `stdout`, and `stderr` arguments of `callr::r_bg()` in `tar_watch()` (@mpadge).
+* Handle `future` errors within targets (@stuvet).
 
 # targets 0.6.0